requirements.txt
django==5.*
django-rest-framework
django-cors-headers
django-filter
mysqlclient